[BE] fix: 템포를 곡 분위기로 변경 및 부족한 로직 추가

See merge request sw_track/class_06/final_project/team09/melody-link!52

+++ server/services/song/songManagementService.js
@@ -4,7 +4,7 @@ const path = require("path");
 
 async function uploadSong({ userId, songInfo, audio, songImage }) {
   // 클라이언트로부터 body로 받아야 할 곡의 정보 (음원, 이미지 제외)
-  const { songName, songDescription, songDuration, songCategory, songTempo } =
+  const { songName, songDescription, songDuration, songCategory, songMood } =
     songInfo;
 
   // 곡을 업로드하는 유저 object id는 song의 songUploader에 저장
@@ -14,7 +14,7 @@ async function uploadSong({ userId, songInfo, audio, songImage }) {
     songUploader: userId,
     songDuration,
     songCategory,
-    songTempo,
+    songMood,
     songImageName: songImage[0].filename,
     songImageLocation: path.join(
       __dirname,
@@ -27,7 +27,7 @@ async function uploadSong({ userId, songInfo, audio, songImage }) {
     ),
   });
 
-  return createSong;
+  return createSong.populate("songUploader");
 }
 
 // songId로 해당하는 곡 찾기
@@ -42,9 +42,12 @@ async function getSongInfo(songId) {
 // songId에 해당하는 곡 수정하기
 async function modifySongInfo({ userId, songId, songInfo, audio, songImage }) {
   const findSong = await Song.findById(songId);
+
+  if (!findSong) return "not-found";
+
   if (findSong.songUploader.toString() !== userId) return "forbidden";
 
-  const { songName, songDescription, songDuration, songCategory, songTempo } =
+  const { songName, songDescription, songDuration, songCategory, songMood } =
     songInfo;
 
   const modifySong = await Song.findByIdAndUpdate(
@@ -54,7 +57,7 @@ async function modifySongInfo({ userId, songId, songInfo, audio, songImage }) {
       songDescription,
       songDuration,
       songCategory,
-      songTempo,
+      songMood,
       songImageName: songImage[0].filename,
       songImageLocation: path.join(
         __dirname,
@@ -70,16 +73,18 @@ async function modifySongInfo({ userId, songId, songInfo, audio, songImage }) {
     { new: true }
   );
 
-  return modifySong;
+  return modifySong.populate("songUploader");
 }
 
 // 곡을 DB에서 삭제
 async function deleteSong(songId, userId) {
   const findSong = await Song.findById(songId);
+  if (!findSong) return "not-found";
+
   if (findSong.songUploader.toString() !== userId) return "forbidden";
 
-  const modifySong = await Song.findByIdAndDelete(songId);
-  return modifySong;
+  const deleteSong = await Song.findByIdAndDelete(songId);
+  return deleteSong;
 }
 
 // 곡 좋아요 누르기 DB에 반영
@@ -87,7 +92,7 @@ async function toggleLike(songId, userId) {
   // 1. songId에 해당하는 유저를 찾아서 그 document의 songLiked 필드에 현재 로그인한 user의 objectId가 있는지를 확인
   // 2. 만약 존재한다면 좋아요를 취소해야 하므로 해당 song의 songLiked 필드에 있는 userId를 삭제하기 + user의 likeSong에서도 songId 삭제하기
   // 3. 만약 존재하지 않는다면 좋아요를 눌러야 하므로 해당 song의 songLiked 필드에 userId를 추가하기 + user의 likeSong에서도 songId 추가하기
-  const findSong = await Song.findById(songId);
+  const findSong = await Song.findById(songId).populate("songUploader");
 
   if (!findSong) return { likeUpdatedSong: null, message: "not-found" };
